+++ Platform.PCL/Realm.Sync.PCL/SyncConfigurationPCL.cs
@@ -55,6 +55,13 @@ namespace Realms.Sync
         }
 
         /// <summary>
+        /// Gets or sets a value indicating whether SSL certificate validation is enabled for the connection associated
+        /// with this configuration value.
+        /// </summary>
+        /// <value><c>true</c> if SSL validation is enabled; otherwise, <c>false</c>. Default value is <c>true</c>.</value>
+        public bool EnableSSLValidation { get; set; }
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="SyncConfiguration"/> class.
         /// </summary>
         /// <param name="user">A valid <see cref="User"/>.</param>